Refactor: ObjectCanvas 컴포넌트 리팩토링, 함수 분리
